feat: improve course content
