+++ src/views/Lesson.vue
@@ -1,0 +1,8 @@
+<template>
+  <CourseDetails></CourseDetails>
+</template>
+
+<script setup>
+import CourseDetails from '@/components/Course/CourseDetails.vue';
+
+</script>
